Create Lesson9 and HomeWork9

+++ HomeWork9/Task3/Program.cs
@@ -1,16 +1,16 @@
 ﻿// Напишите программу вычисления функции Аккермана с помощью рекурсии. Даны два неотрицательных числа m и n.
 // m = 3, n = 2 -> A(m,n) = 29
 
-float InputNumber(string message)
+int InputNumber(string message)
 {
     System.Console.Write(message);
     string readValue = Console.ReadLine();
-    float result = float.Parse(readValue);
+    int result = int.Parse(readValue);
     return result;
 }
 
 
-float FunctionAkker(float m, float n)
+int FunctionAkker(int m, int n)
 {
     if (m == 0)
         return n + 1;
@@ -21,7 +21,7 @@ float FunctionAkker(float m, float n)
 
 void Main () 
 {
-    float res = FunctionAkker(InputNumber("Введите первое число "), InputNumber("Введите второе число "));
+    int res = FunctionAkker(InputNumber("Введите первое число "), InputNumber("Введите второе число "));
     Console.WriteLine(res);
 }
